requirements.txt
cython==0.25.1
flask==0.11.1
kafka-python
